feat: create new action that handles the cloud deletion effects on redux
